extracted all scraping code in another file, cleaned up.
